add NameInput.tsx
